+++ mypy.ini
@@ -1,0 +1,2 @@
+[mypy-googlemaps]
+ignore_missing_imports = True
